requirements.txt
Django==5.0.14
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
celery==5.3.6
redis==4.6.0
django-celery-beat==2.6.0
psycopg2-binary==2.9.11
python-dotenv==1.0.1
requests==2.31.0
python-telegram-bot==20.7
django-filter==23.5
vine==5.1.0
kombu==5.6.2
billiard==4.2.4
python-dateutil==2.9.0.post0